Fixing omz_urlencode:15: command not found: iconv
autoenv calls .autoenv_leave.zsh AFTER leaving the directory, thus the
new cwd is used as .
In my .autoenv_leave.zsh template, ./bin is checked, sometimes
(especially when using `cd $(pipenv --venv)`) this evaluates to true,
i.e. when a bin directory is present in the new cwd; resulting in PATH
being set to the empty string, resulting in other programs not being
found, such as iconv for omz_urlencode. omz_urlencode is thus not the
expected culprit, nor was it a missing iconv – in fact, they were just
the first things to be called after PATH went missing.

Using ${OLDPWD} won't fix it when in a subtree, thus I store the pipenv
--where in a new environment variable to handle this more gracefully.